lms-backend/django-postgres-app/requirements.txt
Django>=4.2,<5.0
psycopg2-binary>=2.9,<3.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
python-dotenv>=1.0,<2.0